Move conversion to mesh module
`mpas_tools.conversion` is still available as an alternative, but
`mpas_tools.mesh.conversion` should be used for clarity whenever
possible.

This will also make work well with `mpas_tools.mesh.creation`, to
be added shortly.
